units in print invoice details

+++ resources/views/dashboard/items/index.blade.php
@@ -51,7 +51,7 @@
                     @if(getRound($amount) != 0)
 
                         <span {{tooltip($unit->name)}}>{{getRound($amount)}}</span>
-                        @php($amount = getFrachtion(\App\Models\ItemStore::where('item_id',$row->id)->sum('amount')))
+                        @php($amount = (float)getFrachtion(\App\Models\ItemStore::where('item_id',$row->id)->sum('amount')))
                     @endif
 
                 </td>
@@ -59,7 +59,7 @@
                 <td>
                     @foreach(\App\Models\Unit::where('item_id',$row->id)->where('ratio','!=',1)->get() as $unit)
 
-                            <span {{tooltip($unit->name)}}>{{getRound(($amount * $unit->ratio))}}</span> @if(($loop->index +1) != \App\Models\Unit::where('item_id',$row->id)->where('ratio','!=',1)->count()) - @endif
+                            <span {{tooltip($unit->name)}}>{{getRound(($amount * (float)$unit->ratio))}}</span> @if(($loop->index +1) != \App\Models\Unit::where('item_id',$row->id)->where('ratio','!=',1)->count()) - @endif
                             @php($amount = getFrachtion(\App\Models\ItemStore::where('item_id',$row->id)->sum('amount')))
 
                     @endforeach
